move TestGetTargetValAmtsForHostZone to a its relative test file
